fixes, fully functionnal
