--- data/hyps/hyp.Objects365.yaml
@@ -1,34 +1,0 @@
-# YOLOv5 🚀 by Ultralytics, GPL-3.0 license
-# Hyperparameters for Objects365 training
-# python train.py --weights yolov5m.pt --data Objects365.yaml --evolve
-# See Hyperparameter Evolution tutorial for details https://github.com/ultralytics/yolov5#tutorials
-
-lr0: 0.00258
-lrf: 0.17
-momentum: 0.779
-weight_decay: 0.00058
-warmup_epochs: 1.33
-warmup_momentum: 0.86
-warmup_bias_lr: 0.0711
-box: 0.0539
-cls: 0.299
-cls_pw: 0.825
-obj: 0.632
-obj_pw: 1.0
-iou_t: 0.2
-anchor_t: 3.44
-anchors: 3.2
-fl_gamma: 0.0
-hsv_h: 0.0188
-hsv_s: 0.704
-hsv_v: 0.36
-degrees: 0.0
-translate: 0.0902
-scale: 0.491
-shear: 0.0
-perspective: 0.0
-flipud: 0.0
-fliplr: 0.5
-mosaic: 1.0
-mixup: 0.0
-copy_paste: 0.0
