+++ utils/datasets/LoadImages.py
@@ -4,7 +4,7 @@ from glob import glob
 import numpy as np, cv2
 # #####################################
 from .utils import IMG_FORMATS, VID_FORMATS
-from utils.augmentations import letterbox
+from .utils import letterbox
 # #####################################
